Cleaning of the project, distribute_Data was moved into tests and some other stuff
